+++ jarvis/modules/audio/voices.py
@@ -62,6 +62,6 @@ def voice_changer(phrase: str = None) -> None:
             voice_default()
             speaker.speak(text=f"Reverting the changes to default voice module {models.env.title}!")
             return
-        elif word_match.word_match(phrase=keyword, match_list=keywords.keywords.ok):
+        elif word_match.word_match(phrase=keyword, match_list=keywords.keywords['ok']):
             speaker.speak(text=random.choice(conversation.acknowledgement))
             return